BibLaTeXML: fix parsing of names given as plain strings
Schema allows <person>string</person> or <author>string</author>:
In the first case if there is no comma in the string, we assume it to be
like a protected string in BibTeX (i.e. between curly braces), otherwise
we parse it. In the second case we take the string "as is".